feat: add CORS middleware to expose HTMX headers
